# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MeterLedger",
    "description": "Usage billing: immutable journal, versioned rates, monthly statements and late-event adjustments. All money is simulated USD; no payment provider is connected.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Operations"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "tags": [
          "Operations"
        ],
        "summary": "Metrics Endpoint",
        "operationId": "metrics_endpoint_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/admin/tenants": {
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Create Tenant",
        "operationId": "create_tenant_admin_tenants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdministratorKey": []
          }
        ]
      }
    },
    "/me": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Me",
        "operationId": "me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "CustomerKey": []
          }
        ]
      }
    },
    "/rates": {
      "get": {
        "tags": [
          "Rates"
        ],
        "summary": "Rates",
        "operationId": "rates_rates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "CustomerKey": []
          }
        ]
      }
    },
    "/admin/rates": {
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Add Rate",
        "operationId": "add_rate_admin_rates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdministratorKey": []
          }
        ]
      }
    },
    "/events": {
      "post": {
        "tags": [
          "Usage"
        ],
        "summary": "Usage",
        "operationId": "usage_events_post",
        "security": [
          {
            "CustomerKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsageInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Usage"
        ],
        "summary": "Events",
        "operationId": "events_events_get",
        "security": [
          {
            "CustomerKey": []
          }
        ],
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/corrections": {
      "post": {
        "tags": [
          "Usage"
        ],
        "summary": "Correction",
        "operationId": "correction_corrections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CustomerKey": []
          }
        ]
      }
    },
    "/billing/{period}/preview": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Preview",
        "operationId": "preview_billing__period__preview_get",
        "security": [
          {
            "CustomerKey": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/billing/issue": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Issue",
        "operationId": "issue_billing_issue_post",
        "security": [
          {
            "CustomerKey": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Get Job",
        "operationId": "get_job_jobs__job_id__get",
        "security": [
          {
            "CustomerKey": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Invoices",
        "operationId": "invoices_invoices_get",
        "security": [
          {
            "CustomerKey": []
          }
        ],
        "parameters": [
          {
            "name": "before_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Get Invoice",
        "operationId": "get_invoice_invoices__invoice_id__get",
        "security": [
          {
            "CustomerKey": []
          }
        ],
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}/verify": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Verify",
        "operationId": "verify_invoices__invoice_id__verify_get",
        "security": [
          {
            "CustomerKey": []
          }
        ],
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}/events": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Invoice Events",
        "operationId": "invoice_events_invoices__invoice_id__events_get",
        "security": [
          {
            "CustomerKey": []
          }
        ],
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/queue": {
      "get": {
        "tags": [
          "Operations"
        ],
        "summary": "Queue Status",
        "operationId": "queue_status_admin_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "AdministratorKey": []
          }
        ]
      }
    },
    "/admin/jobs/{job_id}/retry": {
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Retry Job",
        "operationId": "retry_job_admin_jobs__job_id__retry_post",
        "security": [
          {
            "AdministratorKey": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CorrectionInput": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.:-]+$",
            "title": "Event Id"
          },
          "original_id": {
            "type": "string",
            "format": "uuid",
            "title": "Original Id"
          },
          "quantity": {
            "type": "string",
            "title": "Quantity",
            "decimal_places": 6,
            "gt": 0,
            "le": 1000000000,
            "max_digits": 16
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "original_id",
          "quantity",
          "reason"
        ],
        "title": "CorrectionInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IssueInput": {
        "properties": {
          "period": {
            "type": "string",
            "pattern": "^20\\d{2}-(0[1-9]|1[0-2])$",
            "title": "Period"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "period"
        ],
        "title": "IssueInput"
      },
      "RateInput": {
        "properties": {
          "metric": {
            "type": "string",
            "enum": [
              "api_calls",
              "storage_gb_hours",
              "compute_seconds"
            ],
            "title": "Metric"
          },
          "unit_price": {
            "type": "string",
            "title": "Unit Price",
            "decimal_places": 8,
            "gt": 0,
            "le": 1000000,
            "max_digits": 15
          },
          "effective_at": {
            "type": "string",
            "format": "date-time",
            "title": "Effective At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "metric",
          "unit_price",
          "effective_at"
        ],
        "title": "RateInput"
      },
      "TenantInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TenantInput"
      },
      "UsageInput": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.:-]+$",
            "title": "Event Id"
          },
          "metric": {
            "type": "string",
            "enum": [
              "api_calls",
              "storage_gb_hours",
              "compute_seconds"
            ],
            "title": "Metric"
          },
          "quantity": {
            "type": "string",
            "title": "Quantity",
            "decimal_places": 6,
            "gt": 0,
            "le": 1000000000,
            "max_digits": 16
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "metric",
          "quantity",
          "occurred_at"
        ],
        "title": "UsageInput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "AdministratorKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      },
      "CustomerKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}